dblecpu

+++ Assignment2/ActivationAndLossFunctions.h
@@ -5,41 +5,41 @@
 #include <Eigen/Dense>
 
 //activation functions
-float sigmoid(float x)
+double sigmoid(double x)
 {
-	return 1.0f / 1.0f + exp(-x);
+	return 1.0 / 1.0 + exp(-x);
 }
 
-float sigmoid_prime(float x)
+double sigmoid_prime(double x)
 {
-	float s = sigmoid(x);
+	double s = sigmoid(x);
 	return s * (1 - s);
 }
-float tanh2(float x)
+double tanh2(double x)
 {
 	return tanh(x);
 }
 
-float tanh_prime(float x)
+double tanh_prime(double x)
 {
-	return 1.0f - powf(tanh(x), 2.0f);
+	return 1.0 - powf(tanh(x), 2.0);
 }
 
-float relu(float x)
+double relu(double x)
 {
-	return std::max(x, 0.0f);
+	return std::max(x, 0.0);
 }
-float relu_prime(float x)
+double relu_prime(double x)
 {
-	return (float)((int)(x >= 0));
+	return (double)((int)(x >= 0));
 }
 
-float one_minus(float x)
+double one_minus(double x)
 {
 	return 1 - x;
 }
 //loss function and their derivative
-float mse(Eigen::MatrixXf& y_true, Eigen::MatrixXf& y_pred)
+double mse(Eigen::MatrixXf& y_true, Eigen::MatrixXf& y_pred)
 {
 	auto diff = (y_true - y_pred ).array() ;
 	return  ( diff * diff).mean();
@@ -52,7 +52,7 @@ Eigen::MatrixXf mse_prime(Eigen::MatrixXf& y_true, Eigen::MatrixXf& y_pred)
 }
 
 /*
-float binary_cross_entropy(Eigen::MatrixXf& y_true, Eigen::MatrixXf& y_pred)
+double binary_cross_entropy(Eigen::MatrixXf& y_true, Eigen::MatrixXf& y_pred)
 {
 	return  (-y_true * y_pred.log()).mean() - (y_true.unaryExpr(one_minus)) * (y_pred.unaryExpr(one_minus)).log());
 }
